changes based on first review
